lots of bug fixes

+++ src/models/mhcfovea/predictor.py
@@ -8,7 +8,7 @@ import json
 import time
 import pathlib
 
-from . import BasePredictor
+from . import BasePredictor, SuppressStdout
 
 
 class MHCfoveaPredictor(BasePredictor):
@@ -50,7 +50,7 @@ class MHCfoveaPredictor(BasePredictor):
             log50k = {}
             # peptide should contain none of B, J, O, U, X, Z
             group = group[~group['peptide'].str.contains(r'[BJOUXZ]', regex=True)]
-            group = group[group['peptide'].str.len() <= 16]
+            # group = group[group['peptide'].str.len() <= 16]
             group = group.reset_index(drop=True)
             if len(group) == 0:
                 print(f'No valid peptides for {mhc_name}')
@@ -66,7 +66,8 @@ class MHCfoveaPredictor(BasePredictor):
             input_df.to_csv('peptides.csv', index=False)
                     
             start_time = time.time_ns()
-            run_result = subprocess.run(['env/bin/python', 'mhcfovea/predictor.py', f'{cls._wd}/peptides.csv', 'results'], cwd=cls._exe_dir)
+            with SuppressStdout():
+                run_result = subprocess.run(['env/bin/python', 'mhcfovea/predictor.py', f'{cls._wd}/peptides.csv', 'results'], cwd=cls._exe_dir, stdout=subprocess.DEVNULL)
             end_time = time.time_ns()
             assert run_result.returncode == 0
             times.append(end_time - start_time)
@@ -84,7 +85,7 @@ class MHCfoveaPredictor(BasePredictor):
             grouped_by_len = result_df.groupby(result_df['sequence'].str.len())
             for length, subgroup in grouped_by_len:
                 # print(subgroup.columns)
-                pred[length] = torch.tensor((1 - subgroup['%rank']).tolist(), dtype=torch.double)
+                pred[length] = torch.tensor((100 - subgroup['%rank']).tolist(), dtype=torch.double)
                 label[length] = torch.tensor(subgroup['label'].tolist(), dtype=torch.long)
                 if 'log50k' in subgroup.columns:
                     log50k[length] = torch.tensor(subgroup['log50k'].tolist(), dtype=torch.double)
@@ -106,7 +107,7 @@ class MHCfoveaPredictor(BasePredictor):
         log50ks_diff = {}
 
         for mhc_name, group in df:
-            if not mhc_name.startswith('HLA-'):
+            if not mhc_name.startswith(('HLA-A', 'HLA-B', 'HLA-C')):
                 print(f'Unknown MHC name: {mhc_name}')
                 if cls._unknown_mhc == 'ignore':
                     continue
@@ -117,8 +118,8 @@ class MHCfoveaPredictor(BasePredictor):
             log50k_diff = {}
             group = group[~group['peptide1'].str.contains(r'[BJOUXZ]', regex=True)]
             group = group[~group['peptide2'].str.contains(r'[BJOUXZ]', regex=True)]
-            group = group[group['peptide1'].str.len() <= 16]
-            group = group[group['peptide2'].str.len() <= 16]
+            # group = group[group['peptide1'].str.len() <= 16]
+            # group = group[group['peptide2'].str.len() <= 16]
             group = group.reset_index(drop=True)
             if len(group) == 0:
                 print(f'No valid peptides for {mhc_name}')
@@ -129,8 +130,9 @@ class MHCfoveaPredictor(BasePredictor):
             peptides2 = group['peptide2'].tolist()
             input_df = pd.DataFrame({'sequence': peptides1 + peptides2, 'mhc': [mhc_formatted] * (len(peptides1) + len(peptides2))})
             input_df.to_csv('peptides.csv', index=False)
-
-            run_result = subprocess.run(['env/bin/python', 'mhcfovea/predictor.py', f'{cls._wd}/peptides.csv', 'results'], cwd=cls._exe_dir)
+            
+            with SuppressStdout():
+                run_result = subprocess.run(['env/bin/python', 'mhcfovea/predictor.py', f'{cls._wd}/peptides.csv', 'results'], cwd=cls._exe_dir, stdout=subprocess.DEVNULL)
             assert run_result.returncode == 0
             try:
                 result_df = pd.read_csv(f'{cls._exe_dir}/results/prediction.csv')
@@ -139,13 +141,14 @@ class MHCfoveaPredictor(BasePredictor):
                 print(mhc_name, ' failed')
                 raise e
             
-            result_df1 = result_df.iloc[:len(group)]
-            result_df2 = result_df.iloc[len(group):]
-            parallel_df = pd.DataFrame({'length': group['peptide1'].str.len(), 'pred1': 1 - result_df1['%rank'], 'pred2': 1 - result_df2['%rank'], 'log50k1': group['log50k1'], 'log50k2': group['log50k2']})
-            grouped_by_len = parallel_df.groupby(parallel_df['length'])
+            result_df1 = result_df.iloc[:len(group)].reset_index(drop=True)
+            result_df2 = result_df.iloc[len(group):].reset_index(drop=True)
+            group['pred1'] = 100 - result_df1['%rank']
+            group['pred2'] = 100 - result_df2['%rank']
+            grouped_by_len = group.groupby(group['peptide1'].str.len())
             for length, subgroup in grouped_by_len:
                 pred1 = torch.tensor(subgroup['pred1'].tolist(), dtype=torch.double)
-                pred2 = torch.tensor(subgroup['pred2'], dtype=torch.double)
+                pred2 = torch.tensor(subgroup['pred2'].tolist(), dtype=torch.double)
                 log50k1 = torch.tensor(subgroup['log50k1'].tolist(), dtype=torch.double)
                 log50k2 = torch.tensor(subgroup['log50k2'].tolist(), dtype=torch.double)
                 pred_diff[length] = pred1 - pred2
@@ -154,5 +157,6 @@ class MHCfoveaPredictor(BasePredictor):
         preds_diff[mhc_name] = pred_diff
         log50ks_diff[mhc_name] = log50k_diff
 
-        os.remove('peptides.csv')
+        if os.path.exists('peptides.csv'):
+            os.remove('peptides.csv')
         return (preds_diff,), log50ks_diff
